+++ lib/flutter_ai_toolkit.dart
@@ -14,7 +14,6 @@
 /// - Chat UI: Ready-to-use widgets for displaying chat interfaces.
 library;
 
-export 'src/llm_chat_view_controller.dart';
 export 'src/providers/interface/chat_message.dart';
 export 'src/providers/providers.dart';
 export 'src/styles/styles.dart';