+++ lib/src/providers/interface/chat_message.dart
@@ -2,6 +2,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// using dynamic calls to translate to/from JSON
+// ignore_for_file: avoid_dynamic_calls
+
+import 'dart:convert';
+
 import '../../providers/interface/attachments.dart';
 import 'message_origin.dart';
 
@@ -25,6 +30,37 @@ class ChatMessage {
   })  : _text = text,
         assert(
             origin.isUser && text != null && text.isNotEmpty || origin.isLlm);
+
+  /// Converts a JSON map representation to a [ChatMessage].
+  ///
+  /// The map should contain the following keys:
+  /// - 'origin': The origin of the message (user or model).
+  /// - 'text': The text content of the message.
+  /// - 'attachments': A list of attachments, each represented as a map with:
+  ///   - 'type': The type of the attachment ('file' or 'link').
+  ///   - 'name': The name of the attachment.
+  ///   - 'mimeType': The MIME type of the attachment.
+  ///   - 'data': The data of the attachment, either as a base64 encoded string
+  ///     (for files) or a URL (for links).
+  factory ChatMessage.fromJson(Map<String, dynamic> map) => ChatMessage(
+        origin: MessageOrigin.values.byName(map['origin'] as String),
+        text: map['text'] as String,
+        attachments: [
+          for (final attachment in map['attachments'] as List<dynamic>)
+            switch (attachment['type'] as String) {
+              'file' => FileAttachment.fileOrImage(
+                  name: attachment['name'] as String,
+                  mimeType: attachment['mimeType'] as String,
+                  bytes: base64Decode(attachment['data'] as String),
+                ),
+              'link' => LinkAttachment(
+                  name: attachment['name'] as String,
+                  url: Uri.parse(attachment['data'] as String),
+                ),
+              _ => throw UnimplementedError(),
+            },
+        ],
+      );
 
   /// Factory constructor for creating an LLM-originated message.
   ///
@@ -68,4 +104,38 @@ class ChatMessage {
       'text: $text, '
       'attachments: $attachments'
       ')';
+
+  /// Converts a [ChatMessage] to a JSON map representation.
+  ///
+  /// The map contains the following keys:
+  /// - 'origin': The origin of the message (user or model).
+  /// - 'text': The text content of the message.
+  /// - 'attachments': A list of attachments, each represented as a map with:
+  ///   - 'type': The type of the attachment ('file' or 'link').
+  ///   - 'name': The name of the attachment.
+  ///   - 'mimeType': The MIME type of the attachment.
+  ///   - 'data': The data of the attachment, either as a base64 encoded string
+  ///     (for files) or a URL (for links).
+  Map<String, dynamic> toJson() => {
+        'origin': origin.name,
+        'text': text,
+        'attachments': [
+          for (final attachment in attachments)
+            {
+              'type': switch (attachment) {
+                (FileAttachment _) => 'file',
+                (LinkAttachment _) => 'link',
+              },
+              'name': attachment.name,
+              'mimeType': switch (attachment) {
+                (final FileAttachment a) => a.mimeType,
+                (final LinkAttachment a) => a.mimeType,
+              },
+              'data': switch (attachment) {
+                (final FileAttachment a) => base64Encode(a.bytes),
+                (final LinkAttachment a) => a.url,
+              },
+            },
+        ],
+      };
 }